Removed mythcommflag, ffmpeg system call.  Requires mkvmerge >= 5.0.1
After several weeks of use for exporting 30 min television episodes,
always using the --direct-db-cutlist option, with no ill results, the
"option" has been removed and the direct DB access is not default.
This is advantageous as it removes the mythcommflag system call.

Also, based on a user's input, mkvmerge gained the ability to use
MPEG-TS files as input in version 5.0.0.  These are the files that
MythTV creates from the HDPVR, thus removing the "need" for the ffmpeg
conversion.  I would like to add a version check for future releases
that allows the conditional execution of the ffmpeg command (and, as a
user has indicated, it might be nice to allow ffmpeg to run other
conversions).

Included changelog file.