New dependencies to detect actual frame rate.
The script now also depends on 'mediainfo' in order to return the
actual frame rate of the input video file.  This removes the
hard-coded value of 29.97 that was giving some users errors.

The script also now requires the perl module File::Which to provide a
quick method to search the user's PATH for the external executables,
exiting with a message when they aren't found.